Move RenameTableModal
