+++ src/pages/workspaces/workspace/Data.js
@@ -16,7 +16,6 @@ import EntitiesContent from 'src/components/data/EntitiesContent';
 import ExportDataModal from 'src/components/data/ExportDataModal';
 import FileBrowser from 'src/components/data/FileBrowser';
 import LocalVariablesContent from 'src/components/data/LocalVariablesContent';
-import RenameTableModal from 'src/components/data/RenameTableModal';
 import { useSavedColumnSettings } from 'src/components/data/SavedColumnSettings';
 import WDSContent from 'src/components/data/WDSContent';
 import { WdsTroubleshooter } from 'src/components/data/WdsTroubleshooter';
@@ -26,6 +25,7 @@ import { MenuButton } from 'src/components/MenuButton';
 import { MenuDivider, MenuTrigger } from 'src/components/PopupTrigger';
 import { FlexTable, HeaderCell } from 'src/components/table';
 import { SnapshotInfo } from 'src/components/workspace-utils';
+import { RenameTableModal } from 'src/data/data-table/entity-service/RenameTableModal';
 import { renderDataCell } from 'src/data/data-table/entity-service/renderDataCell';
 import { EntityUploader } from 'src/data/data-table/shared/EntityUploader';
 import { Ajax } from 'src/libs/ajax';
